wf-2: refact: forecast api
