+++ src/main/java/com/cawe/horaperfeita/application/services/WeatherService.java
@@ -1,20 +1,20 @@
 package com.cawe.horaperfeita.application.services;
 
-import com.cawe.horaperfeita.infrastructure.external.client.openMeteor.OpenMeteor;
+import com.cawe.horaperfeita.infrastructure.external.interfaces.IForecast;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class WeatherService {
 
-    private final OpenMeteor openMeteor;
-
-    public WeatherService() {
-        this.openMeteor = new OpenMeteor();
-    }
+    @Autowired
+    private final IForecast iForecast;
 
     public String getSevenDaysForecast(String latitude, String longitude) {
-        this.openMeteor.addTemperature();
-        return this.openMeteor.getWeatherForecastForSevenDays(latitude, longitude);
+        iForecast.addTemperature();
+        iForecast.addLatitudeAndLongitude(latitude, longitude);
+        return this.iForecast.getWeatherForecast();
     }
 }
