+++ src/main/java/com/cawe/horaperfeita/infrastructure/config/ApplicationConfig.java
@@ -1,5 +1,6 @@
 package com.cawe.horaperfeita.infrastructure.config;
 
+import com.cawe.horaperfeita.infrastructure.external.client.openMeteor.OpenMeteor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
@@ -11,4 +12,10 @@ public class ApplicationConfig {
     public RestTemplate restTemplate() {
         return new RestTemplate();
     }
+
+    @Bean
+    public OpenMeteor openMeteor(RestTemplate restTemplate) {
+        return new OpenMeteor(restTemplate);
+    }
+
 }
