--- src/main/java/com/cawe/horaperfeita/infrastructure/external/interfaces/IExternalClient.java
@@ -1,4 +1,0 @@
-package com.cawe.horaperfeita.infrastructure.external.interfaces;
-
-public interface IExternalClient {
-}
