+++ src/main/java/com/cawe/horaperfeita/infrastructure/external/interfaces/IForecast.java
@@ -1,0 +1,9 @@
+package com.cawe.horaperfeita.infrastructure.external.interfaces;
+
+public interface IForecast {
+    IForecast addTemperature();
+
+    IForecast addLatitudeAndLongitude(String latitude, String longitude);
+
+    String getWeatherForecast();
+}
